requirements.txt
altair==4.1.0
ipython==7.24.1
ipython-genutils==0.2.0
ipywidgets==7.6.3
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyter-server==1.13.5
jupyterlab==3.2.9
jupyterlab-pygments==0.1.2
jupyterlab-server==2.10.3
jupyterlab-widgets==1.0.0
matplotlib==3.4.2
matplotlib-inline==0.1.2
notebook==6.4.0
numpy==1.20.3
openpyxl==3.0.7
packaging==20.9
panda==0.3.1
pandas==1.2.4
pandocfilters==1.4.3
parso==0.8.2
patsy==0.5.1
pickleshare==0.7.5
Pillow==8.2.0
pipenv==2021.5.29
platformdirs==2.3.0
plotly==5.1.0
prometheus-client==0.11.0
prompt-toolkit==3.0.19
python-dateutil==2.8.1
requests==2.25.1
scipy==1.7.0
seaborn==0.11.1
streamlit==1.10.0
streamlit-aggrid==0.2.1
yfinance==0.1.59
zipp==3.7.0
tk==0.1.0